+++ back/src/main/java/com/back/domain/order/dto/OrderDetailResponse.java
@@ -1,8 +1,6 @@
 package com.back.domain.order.dto;
 
-import com.back.domain.item.entity.Item;
 import com.back.domain.order.entity.Order;
-import com.back.domain.order.entity.OrderItem;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
@@ -24,18 +22,5 @@ public class OrderDetailResponse {
         this.address = order.getAddress();
         this.postcode = order.getPostcode();
         this.orderItems = orderItems;
-    }
-
-    @Getter
-    public static class OrderItemDto{
-        private final String itemName;
-        private final int count;
-        private final int price;
-
-        public OrderItemDto(Item item, OrderItem orderItem) {
-            this.itemName = item.getItemName();
-            this.count = orderItem.getQuantity();
-            this.price = item.getPrice();
-        }
     }
 }
